Fix gui.
