+++ src/main/java/org/bigtows/window/ui/pinnote/PinNoteComponent.java
@@ -3,14 +3,13 @@ package org.bigtows.window.ui.pinnote;
 import com.intellij.openapi.actionSystem.ActionManager;
 import com.intellij.openapi.actionSystem.ActionToolbar;
 import com.intellij.openapi.actionSystem.DefaultActionGroup;
-import com.intellij.ui.JBColor;
+import com.intellij.ui.components.JBScrollPane;
 import com.intellij.ui.components.JBTabbedPane;
 import com.intellij.util.ui.JBUI;
 import org.bigtows.service.note.notebook.Notebook;
 import org.bigtows.window.ui.border.BottomBorder;
 import org.bigtows.window.ui.notetree.NoteTree;
 import org.bigtows.window.ui.pinnote.action.AddNote;
-import org.bigtows.window.ui.pinnote.action.ForceRefreshNoteAction;
 
 import javax.swing.*;
 import java.awt.*;
@@ -40,7 +39,17 @@ public class PinNoteComponent {
     }
 
     public void addNotebook(Notebook<?> notebook, Icon icon, NoteTree noteTree) {
-        notebookTabbedPane.addTab(notebook.getName(), icon, noteTree);
+        notebookTabbedPane.addTab(
+                notebook.getName(),
+                icon,
+                this.createScrollPaneWithNoteTree(noteTree)
+        );
+    }
+
+    private JScrollPane createScrollPaneWithNoteTree(NoteTree noteTree) {
+        var scroll = new JBScrollPane(noteTree, ScrollPaneConstants.VERTICAL_SCROLLBAR_AS_NEEDED, ScrollPaneConstants.HORIZONTAL_SCROLLBAR_AS_NEEDED);
+        scroll.setBorder(BorderFactory.createEmptyBorder());
+        return scroll;
     }
 }
